project complete

+++ database/migrations/2023_02_22_012106_create_images_table.php
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->string('tracking_no')->nullable();
             $table->tinyInteger('discount')->default('5');
             $table->tinyInteger('tax')->default('13');
-            $table->int('delivery')->default('100');
+            $table->tinyInteger('delivery')->default('100');
             $table->string('message')->nullable();
             $table->timestamps();
         });
